Modified report function for sending flow.
We follow the convention that when a flow is not active, all
metrics (including RTT delay) are reported to be zero. Similarly,
when no packets have been received in an interval, the RTT
delay is reported to be zero.

A potential drawback of this convention is that it may mess
up the overall picture. For instance, if we want to look at average
RTT delay over all flows, the non-active flows would artificially
bring down the average by reporting 0.